evento para borrar imagenes

+++ src/main/java/com/wossha/clothing/commands/removeClothe/RemoveClotheCommand.java
@@ -9,8 +9,8 @@ import com.wossha.clothing.dto.ClotheDTO;
 import com.wossha.clothing.infrastructure.mapper.MapperDozer;
 import com.wossha.clothing.infrastructure.repositories.ClotheRepository;
 import com.wossha.json.events.events.api.Event;
-import com.wossha.json.events.events.pictures.SavePictureEvent.Message;
-import com.wossha.json.events.events.pictures.SavePictureEvent.SavePictureEvent;
+import com.wossha.json.events.events.pictures.RemovePictureEvent.RemovePictureEvent;
+import com.wossha.json.events.events.pictures.RemovePictureEvent.Message;
 import com.wossha.json.events.services.UUIDGenerator;
 import com.wossha.msbase.commands.CommandResult;
 import com.wossha.msbase.commands.ICommand;
@@ -46,14 +46,28 @@ public class RemoveClotheCommand implements ICommand<RemoveClothe> {
 	@Override
 	public CommandResult execute() throws BusinessException, TechnicalException {
 		CommandResult result = new CommandResult();
+		ClotheDTO clothe = repo.findClotheByUuid(data.getUuid());
 		
+		if(clothe == null) {
+			throw new BusinessException("la prenda a eliminar no existe");
+		}
+		
+		Event savePictureEvent = generateRemovePictureEvent(clothe.getPicture());
+		result.addEvent(savePictureEvent);
 		repo.removeClothe(data.getUuid());
+		
 		result.setMessage("La prenda se ha eliminado correctamente");
 
 		
 		return result;
 	}
 
+	
+	private RemovePictureEvent generateRemovePictureEvent(String uuidPicture) {
+		Message message = new Message(uuidPicture);
+		RemovePictureEvent event = new RemovePictureEvent(WosshaClothingApplication.APP_NAME, this.username, message);
+		return event;
+	}
 
 	@Override
 	public void setUsername(String username) {
